Never repeat an ignored job.
This may be a breaking change.

I would like to never repeat a job that is `IgnoreDupe`. Even if the job
is completed.

This just means that the tags have to be globally unique.

For example, if I am fetching URLs, I should include a year/day number,
and then I will never fetch twice *on the same day*.